+++ esibd/devices/lakeshore335/lakeshore.py
@@ -191,7 +191,7 @@ class TemperatureController(DeviceController):
             self.ls335 = Model335(baud_rate=57600, com_port=self.controllerParent.COM)  # may raise AttributeError that can not be excepted
             time.sleep(500)  # TODO test if initialization gets more stable with this delay
             self.signalComm.initCompleteSignal.emit()
-        except (AttributeError, Exception) as e:  # pylint: disable=[broad-except]
+        except (AttributeError, Exception) as e:  # pylint: disable=[broad-except]  # noqa: BLE001
             self.print(f'Error while initializing: {e}', flag=PRINT.ERROR)
         finally:
             self.initializing = False